Fix scraped movie to create input translation
