+++ ui/v2.5/src/components/Scenes/SceneDetails/SceneScrapeDialog.tsx
@@ -22,12 +22,12 @@ import {
   usePerformerCreate,
   useMovieCreate,
   useTagCreate,
-  makePerformerCreateInput,
 } from "src/core/StashService";
 import { useToast } from "src/hooks/Toast";
-import DurationUtils from "src/utils/duration";
 import { useIntl } from "react-intl";
 import { uniq } from "lodash-es";
+import { scrapedPerformerToCreateInput } from "src/core/performers";
+import { scrapedMovieToCreateInput } from "src/core/movies";
 
 interface IScrapedStudioRow {
   title: string;
@@ -464,7 +464,7 @@ export const SceneScrapeDialog: React.FC<ISceneScrapeDialogProps> = ({
   }
 
   async function createNewPerformer(toCreate: GQL.ScrapedPerformer) {
-    const input = makePerformerCreateInput(toCreate);
+    const input = scrapedPerformerToCreateInput(toCreate);
 
     try {
       const result = await createPerformer({
@@ -503,23 +503,8 @@ export const SceneScrapeDialog: React.FC<ISceneScrapeDialogProps> = ({
   }
 
   async function createNewMovie(toCreate: GQL.ScrapedMovie) {
-    let movieInput: GQL.MovieCreateInput = { name: "" };
+    const movieInput = scrapedMovieToCreateInput(toCreate);
     try {
-      movieInput = Object.assign(movieInput, toCreate);
-
-      // #788 - convert duration and rating to the correct type
-      movieInput.duration = DurationUtils.stringToSeconds(
-        toCreate.duration ?? undefined
-      );
-      if (!movieInput.duration) {
-        movieInput.duration = undefined;
-      }
-
-      movieInput.rating = parseInt(toCreate.rating ?? "0", 10);
-      if (!movieInput.rating || Number.isNaN(movieInput.rating)) {
-        movieInput.rating = undefined;
-      }
-
       const result = await createMovie({
         variables: { input: movieInput },
       });
