+++ ui/v2.5/src/core/performers.ts
@@ -1,6 +1,9 @@
 import { PerformersCriterion } from "src/models/list-filter/criteria/performers";
 import * as GQL from "src/core/generated-graphql";
 import { ListFilterModel } from "src/models/list-filter/filter";
+import { stringToGender } from "src/utils/gender";
+import { filterData } from "src/utils/data";
+import { stringToCircumcised } from "src/utils/circumcised";
 
 export const usePerformerFilterHook = (
   performer: GQL.PerformerDataFragment
@@ -79,3 +82,40 @@ export function sortPerformers<T extends IPerformerFragment>(performers: T[]) {
 
   return ret;
 }
+
+export const scrapedPerformerToCreateInput = (
+  toCreate: GQL.ScrapedPerformer
+) => {
+  const input: GQL.PerformerCreateInput = {
+    name: toCreate.name ?? "",
+    url: toCreate.url,
+    gender: stringToGender(toCreate.gender),
+    birthdate: toCreate.birthdate,
+    ethnicity: toCreate.ethnicity,
+    country: toCreate.country,
+    eye_color: toCreate.eye_color,
+    height_cm: toCreate.height ? Number(toCreate.height) : undefined,
+    measurements: toCreate.measurements,
+    fake_tits: toCreate.fake_tits,
+    career_length: toCreate.career_length,
+    tattoos: toCreate.tattoos,
+    piercings: toCreate.piercings,
+    aliases: toCreate.aliases,
+    twitter: toCreate.twitter,
+    instagram: toCreate.instagram,
+    tag_ids: filterData((toCreate.tags ?? []).map((t) => t.stored_id)),
+    image:
+      (toCreate.images ?? []).length > 0
+        ? (toCreate.images ?? [])[0]
+        : undefined,
+    details: toCreate.details,
+    death_date: toCreate.death_date,
+    hair_color: toCreate.hair_color,
+    weight: toCreate.weight ? Number(toCreate.weight) : undefined,
+    penis_length: toCreate.penis_length
+      ? Number(toCreate.penis_length)
+      : undefined,
+    circumcised: stringToCircumcised(toCreate.circumcised),
+  };
+  return input;
+};
